pyproject.toml
[project]
name = "kynesis"
version = "0.2.0"
description = "AI-powered QA evaluation platform — stateless processing API"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "fastapi>=0.115.0",
    "uvicorn>=0.34.0",
    "sqlalchemy[asyncio]>=2.0.0",
    "asyncpg>=0.30.0",
    "python-jose[cryptography]>=3.3.0",
    "pydantic-settings>=2.0.0",
    "httpx>=0.28.0",
    "pydantic-ai",
    "pinecone",
    "logfire[fastapi]",
    "pymupdf",
    "python-docx",
    "langchain-text-splitters",
    "python-multipart",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0.0",
    "pytest-asyncio>=0.24.0",
    "httpx>=0.28.0",
]

[tool.pytest.ini_options]
asyncio_mode = "auto"
